Rename all files with .es6 ext to .js
Closes #32
