+++ engine/matrix/matrix.js
@@ -1,5 +1,5 @@
 
-const utils = require('../utils.es6');
+const utils = require('../utils');
 
 /**
  * Private members
