+++ engine/matrix/svd-golub-reinsch.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Matrix  = require('./matrix.es6');
+const Matrix  = require('./matrix');
 
 function pythag(a, b) {
   var absa = Math.abs(a)
