+++ engine/model/combos.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const utils = require('../utils.es6');
+const utils = require('../utils');
 
 /**
  * Generate all combinations of k terms.
