+++ test/engine/combos.test.js
@@ -4,7 +4,7 @@ const chai    = require('chai')
     , expect  = chai.expect
     , should  = chai.should;
 
-const combos  = require('../../engine/model/combos.es6');
+const combos  = require('../../engine/model/combos');
 
 describe('combos', () => {
