+++ test/engine/model.test.js
@@ -5,7 +5,7 @@ const chai    = require('chai')
     , should  = chai.should;
 
 const Matrix  = require('../../engine/matrix').Matrix
-    , combos  = require('../../engine/model/combos.es6')
+    , combos  = require('../../engine/model/combos')
     , Model   = require('../../engine/model')
     , dataset = require('./test.data.json');
